[fix] Removed the module unloading from the extension quit method
Unloading the modules was blocking the main thread and since modules are unloaded directly by the OS on closing the game we just let the OS deal with it.
If for some weird reason the game as not ended no problem the modules will stay loaded and will be reused normally.